exercicios

+++ estrtuturas de repetição - while e for/exercicio4.py
@@ -1,7 +1,7 @@
 cont = 1 
 a = 0 
 while cont <= 10:
-    numero = int(input("Insira um número: "))
+    numero = float(input("Insira um número: "))
     if numero > 100 and numero < 200:
         a += 1 
     cont += 1
